requirements.txt
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.2
django-heroku==0.3.1
djangorestframework==3.9.0
get==2018.11.19
gunicorn==19.9.0
idna==2.7
mysql-connector-python==8.0.12
numpy==1.15.3
pandas==0.23.4
post==2018.11.20
protobuf==3.6.1
psycopg2==2.7.6.1
public==2018.11.20
python-dateutil==2.7.3
pytz==2018.6
query-string==2018.11.20
request==2018.11.20
requests==2.20.1
scipy==1.1.0
selenium==3.141.0
six==1.11.0
urllib3==1.24.1
whitenoise==4.1.2
xgboost==0.80
Pillow==5.3.0